fix: eliminate int method signatures for TRIC JavaCard converter compatibility
TRIC rejects `int` in method signatures and field descriptors because those
are encoded in the CAP Descriptor component, which standard JavaCard card
runtime cannot handle without the int extension.

Changes:
- MathUtil: inline makeInt/setInt into addInt, xorInt, rotrInt; remove
  the two int-signature helper methods entirely
- NumberUtil: remove byteArrayToInt (returns int) and intToByteArray (takes
  int); all callers (Sha2, Ed25519, ScriptInterpreter) now inline the
  4-byte big-endian read/write directly
- RlpDecoder: change OFFSET_SHORT_LIST and OFFSET_LONG_LIST from int to
  short (values 0xc0/0xf7 fit in short range)
- ScriptInterpreter: change placeholderLength field from int to short

int local variables inside method bodies are unaffected — TRIC handles
those through its I2S (int-to-short) algorithm.

+++ src/coolbitx/RlpDecoder.java
@@ -36,7 +36,7 @@ public class RlpDecoder {
 	 * of the RLP encodings of the items. The range of the first byte is thus
 	 * [0xc0, 0xf7].
 	 */
-	public static int OFFSET_SHORT_LIST = 0xc0;
+	public static short OFFSET_SHORT_LIST = 0xc0;
 
 	/**
 	 * [0xf7] If the total payload of a list is more than 55 bytes long, the RLP
@@ -45,7 +45,7 @@ public class RlpDecoder {
 	 * followed by the concatenation of the RLP encodings of the items. The
 	 * range of the first byte is thus [0xf8, 0xff].
 	 */
-	public static int OFFSET_LONG_LIST = 0xf7;
+	public static short OFFSET_LONG_LIST = 0xf7;
 
 	public static final short RLP_EXCEPTION = 6;
